% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addViewangles.R
\name{addViewangles}
\alias{addViewangles}
\title{Add View Zenith Angle (VZA) and View Azimuth Angle (VAA) to Rays}
\usage{
addViewangles(rays, ...)
}
\arguments{
\item{rays}{An object of class Rays, which contains the trajectory information
of the rays including the X, Y, and Z coordinates of the origin and the target
trajectory points.}
}
\value{
The input Rays object with two new attributes added: VZA and VAA.
}
\description{
Add View Zenith Angle (VZA) and View Azimuth Angle (VAA) to Rays
}
\author{
Benjamin Brede

This function computes the View Zenith Angle (VZA) and View Azimuth Angle (VAA)
for an object of class Rays.
}
