% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin.R
\name{bin}
\alias{bin}
\title{Fast binning of coordinates on a regular grid}
\usage{
bin(x, res, origin = 0)
}
\arguments{
\item{x}{A numeric value representing the coordinate to be binned.}

\item{res}{The resolution of the grid, which determines the spacing between grid points.}

\item{origin}{(optional) The origin point of the grid; defaults to 0 if not specified.}
}
\value{
The binned coordinate value, adjusted to the nearest grid point.
}
\description{
This function takes a numeric value 'x', which represents a coordinate, and
bins it to the nearest grid point based on the specified resolution 'res' and
an optional origin point 'origin'.
}
\author{
Benjamin Brede
}
