% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-Vox.R
\name{Check (Vox)}
\alias{Check (Vox)}
\title{Check if Object is initialized accordingly}
\arguments{
\item{.Object}{Object to check}
}
\description{
Check if Object is initialized accordingly
}
\author{
Benjamin Brede
}
