% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-Rays.R, R/Class-Vox.R
\name{Check <- TODO}
\alias{Check <- TODO}
\title{Check if Object inherits from LAS-class}
\arguments{
\item{.Object}{Object to check}
}
\description{
Check if Object inherits from LAS-class

Check if vox is initilized accordingly
}
\author{
Benjamin Brede
}
