% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clip_rays.R
\name{clip_rays}
\alias{clip_rays}
\title{Fast clipping of rays according to voi}
\usage{
clip_rays(rays, aoi, zrange, buffer, ...)
}
\arguments{
\item{rays}{An object of class 'Rays'.}

\item{aoi}{A numeric vector defining the area of interest with dimensions xmin, ymin, xmax, ymax.}

\item{zrange}{A numeric vector defining the z-range with dimensions zmin, zmax.}

\item{buffer}{numeric of length 3. Buffer to add around voi to avoid edge effects in other processing. Three dimensions as las. Same unit as las}

\item{...}{Additional arguments passed to other methods or functions.}
}
\value{
An object of class 'Rays' that has been clipped according to the specified parameters.
}
\description{
Fast clipping of rays according to voi
}
\details{
Implementation of the Cohen-Sutherland algorithm: \url{https://en.wikipedia.org/wiki/Cohen\%E2\%80\%93Sutherland_algorithm}
}
\author{
Benjamin Brede
}
