% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-Vox.R
\name{head.Vox}
\alias{head.Vox}
\title{Function to display the first few rows of the voxel data contained in a vox object}
\usage{
\method{head}{Vox}(x, ...)
}
\arguments{
\item{x}{a Vox object}

\item{...}{Additional arguments passed to the head function}
}
\value{
A \code{data.frame} showing the first few rows of the voxel data
}
\description{
Function to display the first few rows of the voxel data contained in a vox object
}
\author{
Benjamin Brede
}
