% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-Rays.R
\name{Initialize (Rays)}
\alias{Initialize (Rays)}
\title{Initialize object ray}
\usage{
\S4method{initialize}{Rays}(.Object, ...)
}
\arguments{
\item{.Object}{Object}

\item{...}{Additional arguments passed to other methods or functions.}
}
\value{
a Ray object
}
\description{
Initialize object ray
}
\author{
Benjamin Brede
}
