% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/las2rays.R
\docType{methods}
\name{las2rays}
\alias{las2rays}
\alias{las2rays,LAS,data.table-method}
\alias{las2rays,LAS,numeric-method}
\title{Translate LAS object to Rays}
\usage{
las2rays(las, traj, ...)

\S4method{las2rays}{LAS,data.table}(las, traj)

\S4method{las2rays}{LAS,numeric}(las, traj)
}
\arguments{
\item{las}{A LAS object containing the LiDAR point cloud data.}

\item{traj}{A data table or numeric vector representing the sensor trajectory.
For a data table, it must contain the columns 'Xtraj', 'Ytraj',
'Ztraj', and 'gpstime'. For a numeric vector, it represents the
x, y, and z coordinates of the trajectory.}

\item{...}{Additional arguments passed to other methods or functions.}
}
\value{
Returns a Rays object with the trajectory information included.
}
\description{
This function converts a LAS object, which represents 3D point cloud data,
into a Rays object. The conversion requires a trajectory dataset that contains
the flight path information of the LiDAR sensor.
}
\seealso{
\code{\link{Rays-class}}
}
\author{
Benjamin Brede
}
