% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_tiles.R
\name{prepare_tiles}
\alias{prepare_tiles}
\title{Prepare tiles that define the voxelization region}
\usage{
prepare_tiles(aoi, res, tilesize, ...)
}
\arguments{
\item{aoi}{AOI that defines the voxelization region}

\item{res}{numeric. Resolution for AOI as c(xres, yres)}

\item{tilesize}{size of tiles}

\item{...}{Additional arguments passed to other methods or functions e.g. crs: CRS for tiles as in \code{\link[sf]{st_as_sf}}}
}
\value{
sf-object
}
\description{
This function prepares a grid of tiles based on an area of interest (AOI) for voxelization.
It calculates a bounding box for the AOI and creates a grid of tiles with specified resolution
and tile size, ensuring that the entire AOI is covered.
}
\note{
The function assumes that the input AOI is in the same CRS as specified by the
'crs' parameter.If the CRS does not match, the function may produce incorrect results.
}
\author{
Benjamin Brede
}
