% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-Rays.R
\name{rbind.Rays}
\alias{rbind.Rays}
\title{Combine Rays objects}
\usage{
\method{rbind}{Rays}(...)
}
\arguments{
\item{...}{Additional arguments passed to other methods or functions.}
}
\description{
Combine Rays objects
}
\author{
Benjamin Brede
}
