% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-Rays.R
\name{Show (Rays)}
\alias{Show (Rays)}
\title{Show object Ray}
\usage{
\S4method{show}{Rays}(object)
}
\arguments{
\item{object}{Ray to show}
}
\description{
Show object Ray
}
\note{
copied from lidR/print.R
}
