% SPDX-FileCopyrightText: 2024 Helmholtz Centre Potsdam - GFZ German Research Centre for Geosciences
% SPDX-FileCopyrightText: 2024 Benjamin Brede
% SPDX-License-Identifier: GPL-3.0-or-later

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-Rays.R
\name{show,Rays-method}
\alias{show,Rays-method}
\title{Show object Ray}
\usage{
\S4method{show}{Rays}(object)
}
\arguments{
\item{object}{Ray to show}
}
\description{
Show object Ray
}
\note{
copied from lidR/print.R
}
