% SPDX-FileCopyrightText: 2024 Helmholtz Centre Potsdam - GFZ German Research Centre for Geosciences
% SPDX-FileCopyrightText: 2024 Benjamin Brede
% SPDX-License-Identifier: GPL-3.0-or-later

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-Vox.R
\name{show,Vox-method}
\alias{show,Vox-method}
\title{Show object Vox}
\usage{
\S4method{show}{Vox}(object)
}
\arguments{
\item{object}{Vox to show}
}
\description{
Show object Vox
}
