% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_equal.R
\name{split_equal2}
\alias{split_equal2}
\title{Split vector into equal parts}
\usage{
split_equal2(x, l)
}
\arguments{
\item{x}{vector to split}

\item{l}{length of groups}
}
\value{
list with parts of the vector
}
\description{
This function divides a vector 'x' into groups of length 'l'.
}
\author{
Benjamin Brede
}
