% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitMultiReturn.R
\docType{methods}
\name{splitMultiReturn}
\alias{splitMultiReturn}
\alias{splitMultiReturn,Rays-method}
\title{Split multi-return into single return point cloud}
\usage{
splitMultiReturn(rays, ...)

\S4method{splitMultiReturn}{Rays}(rays)
}
\arguments{
\item{rays}{An object of class 'Rays', which contains the point cloud data with multi-return pulses.}

\item{...}{Additional arguments passed to other methods or functions.}
}
\description{
This function processes a point cloud with multi-return laser pulses, splitting each multi-return
into individual single-return paths. It ensures that each pulse ends at the first return and
initiates a new path from the first to the subsequent returns.
}
\seealso{
\link{las2rays}
}
\author{
Benjamin Brede
}
