% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vox2sds.R
\docType{methods}
\name{vox2sds}
\alias{vox2sds}
\alias{vox2sds,Vox-method}
\title{Translate Vox to sds}
\usage{
vox2sds(vox, ...)

\S4method{vox2sds}{Vox}(vox)
}
\arguments{
\item{vox}{A Vox object.}

\item{...}{Additional arguments passed to other methods or functions.}
}
\value{
A SpatRasterDataset (sds) object.
}
\description{
This function translates a Vox object into a SpatRasterDataset (sds) object.
It rescales the X and Y coordinates to global coordinates for raster creation,
and creates a multi-layer SpatRaster for each feature, where the z dimension is height.
}
\seealso{
\link{sds}
}
\author{
Benjamin Brede
}
