% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vox2transect.R
\docType{methods}
\name{vox2transect}
\alias{vox2transect}
\alias{vox2transect,Vox-method}
\alias{vox2transect,Vox,sf,numeric-method}
\title{Translate Vox to transect}
\usage{
vox2transect(vox, transect, transect_width, ...)

\S4method{vox2transect}{Vox,sf,numeric}(vox, transect, transect_width)
}
\arguments{
\item{vox}{A Vox object.}

\item{transect}{A sf object representing the transect.}

\item{transect_width}{A numeric value representing the width of the transect.}

\item{...}{Additional arguments passed to other methods or functions.}
}
\value{
A SpatRasterDataset (sds) object.
}
\description{
This function translates a Vox object into a transect. It rescales the X and Y coordinates to global coordinates for raster creation,
and creates a multi-layer SpatRaster for each feature, where the z dimension is height.
}
\author{
Benjamin Brede
}
