% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voxelize.R
\name{voxelize,Rays,sf-method}
\alias{voxelize,Rays,sf-method}
\title{Voxelize point cloud with ray-tracing}
\usage{
\S4method{voxelize}{Rays,sf}(
  rays,
  tiles,
  zrange,
  res = c(x = 1, y = 1, z = 1),
  ac_single,
  voxel_mode = c("LAD", "OCC"),
  process_tiles_parallel = 1,
  process_order_tiles = c("random", "seq")
)
}
\arguments{
\item{rays}{An object of class Rays (LAS with added trajectory)}

\item{tiles}{sf. Tiles representing the area to be processed}

\item{zrange}{numeric. Boundaries for volume of interest in vertical dimension c(zmin, zmax)}

\item{res}{numeric. Resolution as single value for all three dimensions or as c(x, y, z)}

\item{ac_single}{numeric. Attenuation coefficient of single vegetation element (\url{https://www.mdpi.com/2072-4292/10/10/1580})}

\item{voxel_mode}{character. "LAD" for Leaf Area Density estimation with various estimators. "OCC" for occlusion mapping.}

\item{process_tiles_parallel}{numeric. Numbers of tiles to process in parallel (only on *nix systems with \link{mclapply})}

\item{process_order_tiles}{character. Processing order of tiles. "random" for randamized processing (typically balances processing of full and empty voxels). "seq" for sequential processing}
}
\value{
Vox object.
}
\description{
Voxelize point cloud with ray-tracing
}
\author{
Benjamin Brede
}
