% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addPulseID.R
\name{addPulseID}
\alias{addPulseID}
\title{Add Pulse ID to LAS Object}
\usage{
addPulseID(las, ...)
}
\arguments{
\item{las}{An object of class LAS with added trajectory}

\item{...}{Additional arguments passed to other methods or functions.}
}
\description{
This function adds a pulse ID attribute to an LAS object that has a trajectory.
}
\author{
Benjamin Brede
}
