% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addViewangles.R
\docType{methods}
\name{addViewangles}
\alias{addViewangles}
\alias{addViewangles,Rays-method}
\title{Add View Zenith Angle (VZA) and View Azimuth Angle (VAA) to Rays}
\usage{
addViewangles(rays, ...)

\S4method{addViewangles}{Rays}(rays)
}
\arguments{
\item{rays}{An object of class Rays, which contains the trajectory information
of the rays including the X, Y, and Z coordinates of the origin and the target
trajectory points.}

\item{...}{Additional arguments passed to other methods or functions.}
}
\value{
The input Rays object with two new attributes added: VZA and VAA.
}
\description{
This function computes the View Zenith Angle (VZA) and View Azimuth Angle (VAA)
for an object of class Rays.
}
\author{
Benjamin Brede
}
