% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-Rays.R
\name{Check (Rays)}
\alias{Check (Rays)}
\title{Check if Object is initialized accordingly and inherits from LAS-Class}
\arguments{
\item{.Object}{Object to check}
}
\description{
Check if Object is initialized accordingly and inherits from LAS-Class
}
\author{
Benjamin Brede
}
