% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-Vox.R
\name{Check <- TODO}
\alias{Check <- TODO}
\title{Check if vox is initilized accordingly}
\arguments{
\item{.Object}{Object to check}
}
\description{
Check if vox is initilized accordingly
}
\author{
Benjamin Brede
}
