% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-Vox.R
\name{head.Vox}
\alias{head.Vox}
\title{Display the First Few Rows of a Vox Object}
\usage{
\method{head}{Vox}(x, ...)
}
\arguments{
\item{x}{A Vox object.}

\item{...}{Additional arguments passed to the \code{head} function.}
}
\value{
A \code{data.frame} showing the first few rows of the voxel data.
}
\description{
Display the First Few Rows of a Vox Object
}
