% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-Rays.R
\name{initialize,Rays-method}
\alias{initialize,Rays-method}
\title{Initialize Object Ray}
\usage{
\S4method{initialize}{Rays}(.Object, ...)
}
\arguments{
\item{.Object}{Object}

\item{...}{Additional arguments passed to other methods or functions.}
}
\value{
a Ray object
}
\description{
Initialize Object Ray
}
\author{
Benjamin Brede
}
