% SPDX-FileCopyrightText: 2024 Helmholtz Centre Potsdam - GFZ German Research Centre for Geosciences
% SPDX-FileCopyrightText: 2024 Benjamin Brede
% SPDX-License-Identifier: GPL-3.0-or-later

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-Rays.R
\name{initialize,Rays-method}
\alias{initialize,Rays-method}
\title{Initialize object ray}
\usage{
\S4method{initialize}{Rays}(.Object, ...)
}
\arguments{
\item{.Object}{Object}

\item{...}{Additional arguments passed to other methods or functions.}
}
\value{
a Ray object
}
\description{
Initialize object ray
}
\author{
Benjamin Brede
}
