% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-Vox.R
\name{initialize,Vox-method}
\alias{initialize,Vox-method}
\title{Initialize object vox}
\usage{
\S4method{initialize}{Vox}(.Object, ...)
}
\arguments{
\item{.Object}{Object}

\item{...}{Additional arguments passed to other methods or functions.}
}
\value{
a vox object
}
\description{
Initialize object vox
}
\author{
Benjamin Brede
}
