% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/las2rays.R
\docType{methods}
\name{las2rays}
\alias{las2rays}
\alias{las2rays,LAS,data.table-method}
\alias{las2rays,LAS,numeric-method}
\title{Translate LAS object to Rays}
\usage{
las2rays(las, traj, ...)

\S4method{las2rays}{LAS,data.table}(las, traj)

\S4method{las2rays}{LAS,numeric}(las, traj)
}
\arguments{
\item{las}{A LAS object containing the LiDAR point cloud data.}

\item{traj}{A data table or numeric vector representing the sensor trajectory.
For a data table, it must contain the columns Xtraj, Ytraj,
Ztraj, and gpstime. For a numeric vector, it represents the
x, y, and z coordinates of the trajectory.}

\item{...}{Additional arguments passed to other methods or functions.}
}
\value{
Returns a Rays object with the trajectory information included.
}
\description{
This function converts a LAS object, which represents 3D point cloud data,
into a Rays object. The conversion requires a trajectory dataset that contains
the flight path information of the LiDAR sensor.
}
\examples{
library(data.table)
library(lidR)
library(sf)
library(dplyr)

# Initialize LAS object
data_file <- system.file("extdata", "H7_LS_F2_H20_200901-120129.laz", package = "voxelizer")
las <- readLAS(data_file)
epsg(las) <- 32631
laz <- las[1:100]

# Create trajectory for the object
traj_file <- system.file("extdata", "H7_LS_F2_H20_200901-120129.traj", package = "voxelizer")
traj <- fread(traj_file, col.names = c('gpstime', 'roll', 'pitch',
                                       'yaw', 'Xorigin', 'Yorigin',
                                       'Zorigin')) \%>\%
  select(gpstime, Xorigin, Yorigin, Zorigin) \%>\%
  rename(Xtraj = Xorigin, Ytraj = Yorigin, Ztraj = Zorigin)

# Create Rays object with las2rays function
rays <- las2rays(laz, traj)

# Further checks in Class-Rays.R

}
\seealso{
\code{\link{Rays-class}}
}
\author{
Benjamin Brede
}
