% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_voxel_height.R
\docType{methods}
\name{normalize_voxel_height}
\alias{normalize_voxel_height}
\alias{normalize_voxel_height,normalize_voxel_height-methods}
\alias{normalize_voxel_height,Vox,SpatRaster-method}
\title{Normalize voxel heights by terrain}
\usage{
normalize_voxel_height(vox, dem, ...)

\S4method{normalize_voxel_height}{Vox,SpatRaster}(vox, dem)
}
\arguments{
\item{vox}{An object of class 'Vox' representing the voxel data.}

\item{dem}{An object of class 'SpatRaster' representing the digital elevation model.}

\item{...}{Additional arguments passed to other methods or functions.}
}
\value{
An object of class 'Vox' with normalized heights.
}
\description{
This function normalizes the heights of voxels based on a digital elevation model (DEM).
It resamples the DEM to match the resolution and extent of the voxels and then adjusts
the voxel heights accordingly.
}
\author{
Benjamin Brede
}
