% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-Rays.R
\docType{class}
\name{Rays-class}
\alias{Rays-class}
\title{An S4 class to represent LiDAR rays}
\description{
Class Rays is the representation of a las/laz file according to the
\href{https://www.asprs.org/wp-content/uploads/2019/07/LAS_1_4_r15.pdf}{LAS file format specifications}. \code{Rays} inherit from \code{LAS} with additional attributes:
\itemize{
\item{\code{Xtraj}, \code{Ytraj}, \code{Ztraj} (numeric): coordinates of the point on the trajectory from where laser pulse was fired}
\item{\code{Xorigin}, \code{Yorigin}, \code{Zorigin} (numeric): coordinates of origin of this segment, e.g. when multi-return pulses are split, or pulses are clipped}
\item{\code{XYZisHit} (logical): if XYZ represents a true hit or virtual end point generated by clipping Rays}}
}
\author{
Benjamin Brede
}
