% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_equal.R
\name{split_equal}
\alias{split_equal}
\title{Split vector into equal parts}
\usage{
split_equal(x, n)
}
\arguments{
\item{x}{vector to split}

\item{n}{number of groups to split x into}
}
\value{
list with parts of the vector
}
\description{
This function divides a vector 'x' into 'n' groups of equal length, or the vector in
groups of length 'l' If 'n' is 1, it returns the original vector. Otherwise, it uses
the 'cut' function to create a factor that splits the sequence along the vector into
'n' groups, which is then used by the 'split' function to divide the vector accordingly.
}
\examples{
# Example 1: Split vector into groups of length 3
split_equal2(1:10, 3)
# Example 2: Split vector into groups of length 5
split_equal2(1:10, 5)
# Example 3: Split vector into groups of length greater than the vector length
split_equal2(1:10, 20)
# Example 4: Split an empty vector into groups of length 3 (returns an empty list)
split_equal2(integer(0), 3)

}
\author{
Benjamin Brede
}
