% SPDX-FileCopyrightText: 2024 Helmholtz Centre Potsdam - GFZ German Research Centre for Geosciences
% SPDX-FileCopyrightText: 2024 Benjamin Brede
% SPDX-License-Identifier: GPL-3.0-or-later

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_equal.R
\name{split_equal2}
\alias{split_equal2}
\title{Split vector into equal parts}
\usage{
split_equal2(x, l)
}
\arguments{
\item{x}{vector to split}

\item{l}{length of groups}
}
\value{
list with parts of the vector
}
\description{
This function divides a vector 'x' into groups of length 'l'.
}
\examples{
# Example 1: Split vector into groups of length 3
split_equal2(1:10, 3)

# Example 2: Split vector into groups of length 5
split_equal2(1:10, 5)

# Example 3: Split vector into groups of length greater than the vector length
split_equal2(1:10, 20)

# Example 4: Split an empty vector into groups of length 3 (returns an empty list)
split_equal2(integer(0), 3)

}
\author{
Benjamin Brede
}
