% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-Vox.R
\docType{class}
\name{Vox-class}
\alias{Vox-class}
\title{An S4 class to represent voxels}
\description{
Class Vox is the representation of a 3D grid of regularly spaced volumes, typically produced by voxelization (\link{voxelize}) of \link{Rays} object.
}
\section{Slots}{

\describe{
\item{\code{data}}{Voxel data with XYZvoxel columns and attributes (depend on mode)}

\item{\code{extent}}{Voxel grid extent in vox crs. (xmin, ymin, zmin, ymin, ymax, zmax)}

\item{\code{resolution}}{Voxel resolution in in vox crs. c(x, y, z)}

\item{\code{mode}}{Voxel mode: LAD (Leaf Area Density metrics) or OCC (occlusion metrics)}

\item{\code{height_normalized}}{Is Zvoxel height normalized?}

\item{\code{crs}}{CRS of Vox as \link{st_crs}}
}}

\author{
Benjamin Brede
}
